# Make some operators and constructors inlineable