--- a/vm/runtime.cpp
+++ b/vm/runtime.cpp
@@ -73,29 +73,6 @@ bool Value::isPointer() const
     }
 }
 
-Value::operator bool () const
-{
-    assert (tag == TAG_BOOL);
-    return word.int64? 1:0;
-}
-
-Value::operator int32_t () const
-{
-    assert (tag == TAG_INT32);
-    return word.int32;
-}
-
-Value::operator float () const
-{
-    assert (tag == TAG_FLOAT32);
-    return word.float32;
-}
-
-Value::operator refptr () const
-{
-    assert (isPointer());
-    return word.ptr;
-}
 
 Value::operator std::string () const
 {