This makes some cast operators and a constructor inlineable, since inlining them makes the interpreter a bit faster.

Before:
```
benchmarks/img_fill.pls               214 ms
benchmarks/incr_field_1m.pls          643 ms
benchmarks/fcalls_10m.zim             989 ms
benchmarks/fib29.pls                  693 ms
benchmarks/loop_cnt_100m.zim         5120 ms
benchmarks/plush_parser.zim           171 ms
benchmarks/saw_wave.pls               541 ms
benchmarks/sine_wave.pls              682 ms
benchmarks/func_audio.pls            2469 ms
--------------------------------------------
geometric mean                        750 ms
```

After:
```
benchmarks/img_fill.pls               210 ms
benchmarks/incr_field_1m.pls          604 ms
benchmarks/fcalls_10m.zim             729 ms
benchmarks/fib29.pls                  615 ms
benchmarks/loop_cnt_100m.zim         2953 ms
benchmarks/plush_parser.zim           164 ms
benchmarks/saw_wave.pls               515 ms
benchmarks/sine_wave.pls              634 ms
benchmarks/func_audio.pls            2268 ms
--------------------------------------------
geometric mean                        649 ms
```

This has a huge impact on loop_cnt_100m, the other benchmarks just show minor improvements.