@tommyettinger 2017-06-19
Better times in all cases? Sounds great to me, even if it's not always much! The code changes look subtle; it's interesting how that can have such a pronounced effect on one of the simplest benchmarks that takes the most time.