@maximecb 2017-06-19
Sidenote: Zeta is now fast enough to do some basic audio synthesis in real time. The last benchmark synthesizes 10 seconds of audio, using lots of function calls.